Add previous() helper to Router.js
This in theory will add the ``previous()`` helper to the router object. In some cases the user (when using inertia.js w/ Laravel) have been required to get the last route and it's params. I realize the "back" button from the browser or the history can be used from the JavaScript API, but I feel having a quicker way to access the data would be a useful addition. 

Note: This wont work immediately as I wrote this via the browser github commit.